+++ src/main/java/org/mattlang/jc/board/bitboard/BitChessBoard.java
@@ -31,6 +31,10 @@ public class BitChessBoard {
         this.pieceBB = pieceBB;
     }
 
+    public long getPieceSet(int pt) {
+        return pieceBB[pt];
+    }
+
     public long getPieceSet(int pt, int color) {
         return pieceBB[pt] & colorBB[color];
     }
@@ -102,6 +106,10 @@ public class BitChessBoard {
         return color == WHITE ? colorBB[nWhite] : colorBB[nBlack];
     }
 
+    public long getColorMask(int color) {
+        return colorBB[color];
+    }
+
     public void set(int i, byte figureCode) {
         long posMask = 1L << i;
 
@@ -130,19 +138,20 @@ public class BitChessBoard {
 
     }
 
-    public void setByMove(int i, byte figureCode) {
+    /**
+     * Set a figure on an EMPTY field.
+     * Note it does not take care to clear the field before, because it assumes it is empty.
+     * the set(..) method on the other hand takes care for that (does a implicit remove of a figure that is on the field
+     * before setting the new figure)
+     * @param i
+     * @param figureCode
+     */
+    public void setOnEmptyField(int i, byte figureCode) {
         long posMask = 1L << i;
-
-        long invPosMask = ~posMask;
-
         int colorIdx = ((figureCode & BLACK.code) == BLACK.code) ? nBlack : nWhite;
-        int otherColor = colorIdx == nWhite ? nBlack : nWhite;
         byte figType = (byte) (figureCode & MASK_OUT_COLOR);
         colorBB[colorIdx] |= posMask;
-        colorBB[otherColor] &= invPosMask;
-
         pieceBB[figType] |= posMask;
-
     }
 
     public void cleanPos(int i) {
@@ -219,9 +228,7 @@ public class BitChessBoard {
 
     }
 
-    public void move(int from, int to, byte figureCode, int color, byte capturedPiece) {
-
-        byte figType = (byte) (figureCode & MASK_OUT_COLOR);
+    public void move(int from, int to, byte figType, int color, byte capturedPiece) {
 
         long fromBB = 1L << from;
         long toBB = 1L << to;