+++ src/main/java/org/mattlang/jc/engine/evaluation/minimalpst/MinimalPstEvaluation.java
@@ -5,7 +5,6 @@ import static org.mattlang.jc.board.bitboard.BitChessBoard.nWhite;
 
 import org.mattlang.jc.board.BoardRepresentation;
 import org.mattlang.jc.board.Color;
-import org.mattlang.jc.board.PieceList;
 import org.mattlang.jc.board.bitboard.BitBoard;
 import org.mattlang.jc.board.bitboard.BitChessBoard;
 import org.mattlang.jc.engine.EvaluateFunction;
@@ -156,30 +155,7 @@ public class MinimalPstEvaluation implements EvaluateFunction {
 
     @Override
     public int eval(BoardRepresentation currBoard, Color who2Move) {
-
-        if (currBoard instanceof BitBoard) {
-            return evalBitboard((BitBoard) currBoard, who2Move);
-        } else {
-            int who2mov = who2Move == Color.WHITE ? 1 : -1;
-            PieceList wp = currBoard.getWhitePieces();
-            PieceList bp = currBoard.getBlackPieces();
-
-            int midGame = PAWN_MG.calcScore(wp.getPawns(), bp.getPawns(), who2mov) +
-                    KNIGHT_MG.calcScore(wp.getKnights(), bp.getKnights(), who2mov) +
-                    BISHOP_MG.calcScore(wp.getBishops(), bp.getBishops(), who2mov) +
-                    ROOK_MG.calcScore(wp.getRooks(), bp.getRooks(), who2mov) +
-                    QUEEN_MG.calcScore(wp.getQueens(), bp.getQueens(), who2mov) +
-                    KING_MG.calcScore(wp.getKing(), bp.getKing(), who2mov);
-            int endGame = PAWN_EG.calcScore(wp.getPawns(), bp.getPawns(), who2mov) +
-                    KNIGHT_EG.calcScore(wp.getKnights(), bp.getKnights(), who2mov) +
-                    BISHOP_EG.calcScore(wp.getBishops(), bp.getBishops(), who2mov) +
-                    ROOK_EG.calcScore(wp.getRooks(), bp.getRooks(), who2mov) +
-                    QUEEN_EG.calcScore(wp.getQueens(), bp.getQueens(), who2mov) +
-                    KING_EG.calcScore(wp.getKing(), bp.getKing(), who2mov);
-
-            double score = PhaseCalculator.scaleByPhase(wp, bp, midGame, endGame);
-            return (int) score;
-        }
+        return evalBitboard((BitBoard) currBoard, who2Move);
     }
 
     private int evalBitboard(BitBoard currBoard, Color who2Move) {